Fix long double to double conversion warning. JIRA: CXX-14376
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@106444 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/stdtypes.cpp
@@ -240,7 +240,7 @@ bool CPrimitiveTypeFunctions<long double>::Equals(TConstObjectPtr obj1,
         return false;
     }
     // We use DBL_EPSILON because I/O is double-based anyway.
-    return (x == y  ||  fabs(x - y) < fabs(x + y) * DBL_EPSILON || x_Practically_Equal<double, Uint8>(x,y));
+    return (x == y  ||  fabs(x - y) < fabs(x + y) * DBL_EPSILON || x_Practically_Equal<double, Uint8>(double(x),double(y)));
 }
 #endif
